+++ card-management-service/servico-cartao/src/main/java/com/github/joanerson/servico_cartao/cartao/application/service/CartaoService.java
@@ -1,5 +1,6 @@
 package com.github.joanerson.servico_cartao.cartao.application.service;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import com.github.joanerson.servico_cartao.cartao.application.api.request.CartaoNovoRequest;
@@ -8,6 +9,6 @@ import com.github.joanerson.servico_cartao.cartao.application.api.response.Clien
 
 public interface CartaoService {
 	void criaNovoCartao(CartaoNovoRequest cartaoRequest);
-	List<CartaoResponse> buscaCartoesRendaAte(Long renda);
+	List<CartaoResponse> buscaCartoesRendaAte(BigDecimal renda);
 	List<ClienteCartaoResponse> buscaCartaoPorCpf(String cpf);
 }
